[closes #95, #128] Change Type of variables / Remove unnecessary type casting (Part3)

+++ kernel-rs/src/kalloc.rs
@@ -66,7 +66,7 @@ pub unsafe fn freerange(pa_start: *mut libc::CVoid, pa_end: *mut libc::CVoid) {
 /// call to kalloc().  (The exception is when
 /// initializing the allocator; see kinit above.)
 pub unsafe fn kfree(pa: *mut libc::CVoid) {
-    if (pa as usize).wrapping_rem(PGSIZE as usize) != 0
+    if (pa as usize).wrapping_rem(PGSIZE) != 0
         || (pa as *mut u8) < end.as_mut_ptr()
         || pa as usize >= PHYSTOP
     {
@@ -74,7 +74,7 @@ pub unsafe fn kfree(pa: *mut libc::CVoid) {
     }
 
     // Fill with junk to catch dangling refs.
-    ptr::write_bytes(pa as *mut libc::CVoid, 1, PGSIZE as usize);
+    ptr::write_bytes(pa as *mut libc::CVoid, 1, PGSIZE);
     let mut r: *mut Run = pa as *mut Run;
     KMEM.lock.acquire();
     (*r).next = KMEM.freelist;
@@ -94,7 +94,7 @@ pub unsafe fn kalloc() -> *mut libc::CVoid {
     KMEM.lock.release();
     if !r.is_null() {
         // fill with junk
-        ptr::write_bytes(r as *mut u8 as *mut libc::CVoid, 5, PGSIZE as usize);
+        ptr::write_bytes(r as *mut u8 as *mut libc::CVoid, 5, PGSIZE);
     }
     r as *mut libc::CVoid
 }
